Moved threadContext helper functions to threadContext.ts
